Added implementation for Apollo iOS
