--- androidApp/src/main/java/fr/paug/androidmakers/util/Images.kt
@@ -1,15 +1,0 @@
-package fr.paug.androidmakers.util
-
-internal fun imageUrl(relativeUrl: String): String {
-  return "https://raw.githubusercontent.com/paug/android-makers-2022/main/$relativeUrl"
-      .replace("..", "")
-      .replace(".svg", ".svg.png")
-      .let {
-        if (it.endsWith(".svg.png")) {
-          it.replace("logos/", "logos/pngs/")
-        } else {
-          it
-        }
-      }
-
-}
